Refactoring && Tests
